# Optimize compression by avoiding unpredictable branches