@Cyan4973 2024-09-24
I had a look into this, and unfortunately observed a performance regression with this patch on the test system: 

core i7-9700k (noturbo), Ubuntu 24.04, gcc 13.2, `silesia.tar`:
| level | `dev` | this `PR` | difference | comment
| --- | --- | --- | --- | --- |
| -3 | 493 | 484 | -9 | 
| -2 | 466 | 456 | -10 | 
| -1 | 421 | 403 | -18 |
| 1 | 386 | 371 | -15 |
| 2 | 290 | 284 | -6 | difference is lowered when data access pattern reaches L3 cache
| 3 | 208 | 202 | -6 |
| 4 | 196 | 196 | 0 | should be impacted by this patch but is not, possibly due to waiting for data from L3 cache
| 5 | 115 | 115 | 0 | control, should not be impacted

It might be one of these cases where the performance difference is more impacted by random variation in instruction alignment, making it difficult to determine if an observed performance difference is really related to the patch, or a random fluke.

To mitigate the impact of a specific compiler version, I'll try to collect the performance differences on a larger set of compilers and versions.