@danlark1 2024-09-24
Also note gcc vs clang. At Google we exclusively use clang