@TocarIP 2024-09-24
Right, also it is a rolling release model, so doesn't directly corresponds to clang/llvm releases.