@Cyan4973 2024-09-24
I have more results, using a larger set of compilers.
Focusing the benchmarking on level 1, which seems a good data point to capture the impact of this PR.

core i7-9700k (noturbo), Ubuntu 24.04, silesia.tar, level 1:

| compiler | `dev` | this `PR` | difference |
| --- | --- | --- | --- |
| `gcc-9` | 378 | 372 | -6
|`gcc-10` | 381 | 370 | -11 
|`gcc-11` | 379 | 372 | -7 
|`gcc-12` | 385 | 368 | -17
|`gcc-13` | 390 | 367 | -23
|`clang-14`| 379 | 365 | -14
|`clang-15`| 372 | 354 | -18
|`clang-16`| 380 | 367 | -13
|`clang-17`| 376 | 373 | -3
|`clang-18`| 381 | 364 | -17

Unfortunately, this looks like a pretty consistent loss, across all compilers, including `clang`.

Not sure what could explain such an observation difference,
I doubt the performance benefit depends on some novelty in `llvm-trunk`,
so this could be due to differences in compiler settings or hardware ?
The i7-9700k is a Skylake-era cpu, a well-trodden architecture with variants deployed over a fairly large portion of active datacenters.