@danlark1 2024-09-24
On clang-17 I have anectodal improvement after running some cli and perf stat commands. TL;DR some speed improvement and 1.5% fewer branch mispredictions

Overall I used `./zstd --zstd=wlog=15 -b1  ~/silesia/silesia.tar`. If I don't specify the wlog, br misprediction rate is the same

```
 danilak @ danilak1 in ~/zstd/build/build_new (dev …1) 0 [23:20:02]
~ ../build_old/programs/zstd --zstd=wlog=15 -b1  ~/silesia/silesia.tar
 1#silesia.tar       : 211957760 ->  79871780 (x2.654),  355.0 MB/s, 1353.0 MB/s

& danilak @ danilak1 in ~/zstd/build/build_new (dev …1) 0 [23:20:56]
~ programs/zstd --zstd=wlog=15 -b1  ~/silesia/silesia.tar
 1#silesia.tar       : 211957760 ->  79871780 (x2.654),  398.9 MB/s, 1110.8 MB/s

& danilak @ danilak1 in ~/zstd/build/build_new (dev …1) 0 [23:21:12]
~ ../build_old/programs/zstd --zstd=wlog=15 -b1  ~/silesia/silesia.tar
 1#silesia.tar       : 211957760 ->  79871780 (x2.654),  349.0 MB/s, 1298.7 MB/s

& danilak @ danilak1 in ~/zstd/build/build_new (dev …1) 0 [23:21:23]
~ programs/zstd --zstd=wlog=15 -b1  ~/silesia/silesia.tar
 1#silesia.tar       : 211957760 ->  79871780 (x2.654),  393.6 MB/s, 1011.0 MB/s

& danilak @ danilak1 in ~/zstd/build/build_new (dev …1) 0 [23:21:32]
~ sudo perf stat programs/zstd --zstd=wlog=15 -b1  ~/silesia/silesia.tar
 1#silesia.tar       : 211957760 ->  79871780 (x2.654),  457.0 MB/s, 1357.5 MB/s

 Performance counter stats for 'programs/zstd --zstd=wlog=15 -b1 /home/danilak/silesia/silesia.tar':

          7,813.00 msec task-clock                       #    1.000 CPUs utilized             
               149      context-switches                 #   19.071 /sec                      
                 5      cpu-migrations                   #    0.640 /sec                      
               745      page-faults                      #   95.354 /sec                      
    28,338,622,044      cycles                           #    3.627 GHz                       
    78,055,510,248      instructions                     #    2.75  insn per cycle            
     7,438,789,171      branches                         #  952.104 M/sec                     
       220,283,420      branch-misses                    #    2.96% of all branches           
                        TopdownL1                 #     12.5 %  tma_backend_bound      
                                                  #     37.4 %  tma_bad_speculation    
                                                  #      4.5 %  tma_frontend_bound     
                                                  #     45.6 %  tma_retiring           

       7.816908293 seconds time elapsed

       7.725589000 seconds user
       0.087972000 seconds sys



& danilak @ danilak1 in ~/zstd/build/build_new (dev …1) 0 [23:21:43]
~ sudo perf stat ../build_old/programs/zstd --zstd=wlog=15 -b1  ~/silesia/silesia.tar
1#silesia.tar       : 211957760 ->  79871780 (x2.654),  354.0 MB/s, 1389.2 MB/s

 Performance counter stats for '../build/programs/zstd --zstd=wlog=15 -b1 /home/danilak/silesia/silesia.tar':

          6,959.95 msec task-clock                       #    1.000 CPUs utilized             
                83      context-switches                 #   11.925 /sec                      
                 4      cpu-migrations                   #    0.575 /sec                      
               741      page-faults                      #  106.466 /sec                      
    29,116,934,179      cycles                           #    4.183 GHz                       
    71,758,857,772      instructions                     #    2.46  insn per cycle            
     7,105,568,042      branches                         #    1.021 G/sec                     
       324,577,377      branch-misses                    #    4.57% of all branches           
                        TopdownL1                 #     13.2 %  tma_backend_bound      
                                                  #     42.5 %  tma_bad_speculation    
                                                  #      4.8 %  tma_frontend_bound     
                                                  #     39.4 %  tma_retiring           

       6.962525456 seconds time elapsed

       6.856733000 seconds user
       0.104011000 seconds sys
```