@TocarIP 2024-09-24
I'm getting similar results on a skylake (server/worksation variant) with clang 16
```
cat /proc/cpuinfo
...
model name	: Intel(R) Xeon(R) W-2135 CPU @ 3.70GHz
...
```
With `./programs/zstd  -b2 --zstd=hashLog=16,windowLog=16 -c -2 silesia.tar` branch misprediction (as measured by `perf stat -ddd`, so combined for compress+decompress) goes from  4.4% mispredict to 2.9% and speed goes from 370MB/s to 420MB/s.
Without extra parameters speed is slightly worse 395 vs 400 MB/s. I'll take a look at level1 where the regression is worse